Russell says there's a reason at least the int types were explicitly cast to long (see commit log for 271ff6421cb1e3feb12c1955c58c0ec098cc261c), so restore those casts (but leave the float types alone).

+++ tests/testTask.py
@@ -200,13 +200,13 @@ class TaskTestCase(unittest.TestCase):
                              ("CpuTime", float),
                              ("UserTime", float),
                              ("SystemTime", float),
-                             ("MaxResidentSetSize", int),
-                             ("MinorPageFaults", int),
-                             ("MajorPageFaults", int),
-                             ("BlockInputs", int),
-                             ("BlockOutputs", int),
-                             ("VoluntaryContextSwitches", int),
-                             ("InvoluntaryContextSwitches", int),
+                             ("MaxResidentSetSize", long),
+                             ("MinorPageFaults", long),
+                             ("MajorPageFaults", long),
+                             ("BlockInputs", long),
+                             ("BlockOutputs", long),
+                             ("VoluntaryContextSwitches", long),
+                             ("InvoluntaryContextSwitches", long),
                              ):
             for when in ("Start", "End"):
                 for method in ("run", "context"):
